# added performance benchmarks for FSM