Fixes #2560

## Changes

added performance benchmarks for FSM

## Checklist

For significant changes, please ensure that the following have been completed (delete if not relevant):

* [x] This change follows the [Akka.NET API Compatibility Guidelines](https://getakka.net/community/contributing/api-changes-compatibility.html).
* [x] This change follows the [Akka.NET Wire Compatibility Guidelines](https://getakka.net/community/contributing/wire-compatibility.html).
* [x] I have [reviewed my own pull request](https://getakka.net/community/contributing/index.html#review-your-own-pull-requests).
* [x] Design discussion issue #2560 

### This PR's Benchmarks

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19044.2006 (21H2)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=6.0.201
  [Host]     : .NET 6.0.3 (6.0.322.12309), X64 RyuJIT
  DefaultJob : .NET 6.0.3 (6.0.322.12309), X64 RyuJIT


```
|           Method | MsgCount |     Mean |    Error |   StdDev | Ratio | RatioSD |      Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|----------------- |--------- |---------:|---------:|---------:|------:|--------:|-----------:|----------:|----------:|----------:|
|     BenchmarkFsm |  1000000 | 758.8 ms | 10.92 ms | 10.21 ms |  1.67 |    0.03 | 69000.0000 | 4000.0000 | 1000.0000 |    287 MB |
| BenchmarkUntyped |  1000000 | 456.2 ms |  8.86 ms | 10.21 ms |  1.00 |    0.00 | 11000.0000 | 3000.0000 | 1000.0000 |     56 MB |

LOTS of memory allocations for FSMs, and a 67% perf penalty.